refactor: extract recipe service and frontend api wrapper
Backend: move DB logic out of routes/recipes.js into services/recipes.js
so route handlers only deal with HTTP concerns. No behavior change.

Frontend: add src/api/recipes.js wrapper around fetch; pages and
CountryPanel call named functions instead of building URLs inline.
